+++ package/ui/components/button/Button.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styled from "styled-components";
+import styled, { css } from "styled-components";
 
 interface ButtonStyledProps {
   /**
