+++ package/ui/components/index.ts
@@ -1,5 +1,5 @@
-export { default as Button } from "./Button";
-export { default as Template } from "./Template";
-export { default as RegisterTemplate } from "./RegisterTemplate";
-export { default as Portal } from "./Portal";
-export { default as ModalTemplate } from "./ModalTemplate";
+export { default as Button } from "./button/Button";
+export { default as Template } from "./template/Template";
+export { default as RegisterTemplate } from "./template/RegisterTemplate";
+export { default as Portal } from "./modal/Portal";
+export { default as ModalTemplate } from "./modal/ModalTemplate";
