+++ package/ui/components/modal/ModalTemplate.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { transitions } from "../styles";
+import { transitions } from "../../styles";
 import Portal from "./Portal";
 
 interface ModalInnerStyled {
